# Conflicts:
#	core/src/mindustry/client/Client.kt
#	core/src/mindustry/client/ClientVars.java
#	core/src/mindustry/core/NetClient.java

+++ core/src/mindustry/game/Saves.java
@@ -1,12 +1,11 @@
 package mindustry.game;
 
 import arc.*;
-import arc.assets.*;
 import arc.files.*;
 import arc.graphics.*;
+import arc.graphics.g2d.*;
 import arc.struct.*;
 import arc.util.*;
-import arc.util.async.*;
 import mindustry.*;
 import mindustry.core.GameState.*;
 import mindustry.game.EventType.*;
@@ -18,23 +17,26 @@ import mindustry.type.*;
 import java.io.*;
 import java.text.*;
 import java.util.*;
+import java.util.concurrent.*;
 
 import static mindustry.Vars.*;
 
 public class Saves{
-    Seq<SaveSlot> saves = new Seq<>();
+    private static final DateFormat dateFormat = SimpleDateFormat.getDateTimeInstance();
+
+    final Seq<SaveSlot> saves = new Seq<>(0);
     @Nullable SaveSlot current;
     private @Nullable SaveSlot lastSectorSave;
-    AsyncExecutor previewExecutor = new AsyncExecutor(1);
     private boolean saving;
     private float time;
+    public boolean hasLoaded;
 
     long totalPlaytime;
     private long lastTimestamp;
 
-    public Saves(){
-        Core.assets.setLoader(Texture.class, ".spreview", new SavePreviewLoader());
+    private final LinkedBlockingDeque<Runnable> previewQueue = new LinkedBlockingDeque<>();
 
+    public Saves(){
         Events.on(StateChangeEvent.class, event -> {
             if(event.to == State.menu){
                 totalPlaytime = 0;
@@ -44,30 +46,51 @@ public class Saves{
         });
     }
 
+    /** Loads all saves */
     public void load(){
-        saves.clear();
+        load(false);
+    }
 
-        for(Fi file : saveDirectory.list()){
-            if(!file.name().contains("backup") && SaveIO.isSaveValid(file)){
-                SaveSlot slot = new SaveSlot(file);
-                saves.add(slot);
-                slot.meta = SaveIO.getMeta(file);
-            }
-        }
+    /** Loads all saves (unless sectorsOnly is specified in which case only sectors are loaded in order to make campaign load faster). */
+    public void load(boolean sectorsOnly){
+        Time.mark(); Time.mark();
+        hasLoaded = true;
 
-        //clear saves from build <130 that had the new naval sectors.
-        saves.removeAll(s -> {
-            if(s.getSector() != null && (s.getSector().id == 108 || s.getSector().id == 216) && s.meta.build <= 130 && s.meta.build > 0){
-                s.getSector().clearInfo();
-                s.file.delete();
-                return true;
-            }
-            return false;
+        var tasks = new Seq<Future<SaveSlot>>();
+        saveDirectory.walk(file -> {
+            var name = file.name();
+            if (name.endsWith("backup.msav") || !name.endsWith(".msav") || sectorsOnly && !name.startsWith("sector-")) return;
+            tasks.add(mainExecutor.submit(() -> {
+                try{
+                    var s = new SaveSlot(file, SaveIO.getMeta(file));
+                    //clear saves from build <130 that had the new naval sectors.
+                    if(s.getSector() != null && (s.getSector().id == 108 || s.getSector().id == 216) && s.meta.build <= 130 && s.meta.build > 0){
+                        s.getSector().clearInfo();
+                        s.file.delete();
+                    }
+                    return s;
+                }catch(Throwable e){
+                    Log.err("Failed to load save '" + file.name() + "'", e);
+                    return null;
+                }
+            }));
         });
+        var queued = Time.elapsed();
 
+        Time.mark();
+        saves.ensureCapacity(tasks.size);
+        for(Future<SaveSlot> task : tasks){
+            var s = Threads.await(task);
+            if(s != null) saves.add(s);
+        }
+        var blocked = Time.elapsed();
+        var loaded = Time.elapsed();
+
+        Time.mark();
         lastSectorSave = saves.find(s -> s.isSector() && s.getName().equals(Core.settings.getString("last-sector-save", "<none>")));
 
-        //automatically assign sector save slots
+        //automatically (re)assign sector save slots
+        content.planets().each(p -> p.sectors.each(s -> s.save = null)); // FINISHME: This is most likely a horrible idea
         for(SaveSlot slot : saves){
             if(slot.getSector() != null){
                 if(slot.getSector().save != null){
@@ -76,6 +99,14 @@ public class Saves{
                 slot.getSector().save = slot;
             }
         }
+
+        Log.debug("Queued saves in: @ms | Loaded @ saves in: @ms | Blocked for: @ms | Post processed saves in in @ms", queued, saves.size, loaded, blocked, Time.elapsed());
+    }
+
+    /** Unload all saves to reclaim resources */
+    public void unload(){
+        saves.each(SaveSlot::unloadTexture);
+        saves.clear().shrink(); // Don't want all of this stuff in memory
     }
 
     public @Nullable SaveSlot getLastSector(){
@@ -93,6 +124,11 @@ public class Saves{
                 totalPlaytime += Time.timeSinceMillis(lastTimestamp);
             }
             lastTimestamp = Time.millis();
+        }
+
+        var start = Time.millis();
+        while (previewQueue.size() > 0 && Time.timeSinceMillis(start) < 15) {
+            previewQueue.pop().run();
         }
 
         if(state.isGame() && !state.gameOver && current != null && current.isAutosave()){
@@ -136,6 +172,7 @@ public class Saves{
             sector.save = new SaveSlot(getSectorFile(sector));
             sector.save.setName(sector.save.file.nameWithoutExtension());
             saves.add(sector.save);
+            if (saves.size == 1) unload();
         }
         sector.save.setAutosave(true);
         sector.save.save();
@@ -147,6 +184,7 @@ public class Saves{
         SaveSlot slot = new SaveSlot(getNextSlotFile());
         slot.setName(name);
         saves.add(slot);
+        if (saves.size == 1) unload();
         slot.save();
         return slot;
     }
@@ -157,6 +195,7 @@ public class Saves{
         slot.setName(file.nameWithoutExtension());
 
         saves.add(slot);
+        if (saves.size == 1) unload();
         slot.meta = SaveIO.getMeta(slot.file);
         current = slot;
         return slot;
@@ -172,24 +211,33 @@ public class Saves{
     }
 
     public Seq<SaveSlot> getSaveSlots(){
+        if(saves.isEmpty()) load();
         return saves;
     }
 
+    public int saveCount(){
+        return saveDirectory.findAll(f -> !f.name().contains("backup") && f.extEquals("msav")).size;
+    }
+
     public void deleteAll(){
-        for(SaveSlot slot : saves.copy()){
-            if(!slot.isSector()){
-                slot.delete();
-            }
-        }
+        var needsLoad = saves.isEmpty();
+        if(needsLoad) load(); // Need to load them in order to delete them.
+        saves.filter(s -> !s.isSector()).each(SaveSlot::delete); // Delete non sectors
+        if(needsLoad) unload(); // Unload if we just loaded
     }
 
     public class SaveSlot{
         public final Fi file;
-        boolean requestedPreview;
+        private volatile TextureRegion preview;
         public SaveMeta meta;
 
         public SaveSlot(Fi file){
+            this(file, null);
+        }
+
+        public SaveSlot(Fi file, SaveMeta meta){
             this.file = file;
+            this.meta = meta;
         }
 
         public void load() throws SaveException{
@@ -205,9 +253,7 @@ public class Saves{
         }
 
         public void save(){
-            long time = totalPlaytime;
             long prev = totalPlaytime;
-            totalPlaytime = time;
 
             SaveIO.save(file);
             meta = SaveIO.getMeta(file);
@@ -220,29 +266,41 @@ public class Saves{
         }
 
         private void savePreview(){
-            if(Core.assets.isLoaded(loadPreviewFile().path())){
-                Core.assets.unload(loadPreviewFile().path());
-            }
-            previewExecutor.submit(() -> {
+            mainExecutor.submit(() -> {
                 try{
                     previewFile().writePng(renderer.minimap.getPixmap());
-                    requestedPreview = false;
+                    previewQueue.add(this::unloadTexture);
                 }catch(Throwable t){
                     Log.err(t);
                 }
             });
         }
 
-        public Texture previewTexture(){
-            if(!previewFile().exists()){
-                return null;
-            }else if(Core.assets.isLoaded(loadPreviewFile().path())){
-                return Core.assets.get(loadPreviewFile().path());
-            }else if(!requestedPreview){
-                Core.assets.load(new AssetDescriptor<>(loadPreviewFile(), Texture.class));
-                requestedPreview = true;
+        /** Asynchronously loads this save's preview on demand */
+        public TextureRegion previewTexture(){
+            if(preview == null){
+                preview = Core.atlas.find("nomap"); // Prevents loading twice
+                if(previewFile().exists()){
+                    mainExecutor.execute(() -> {
+                        if (preview == null) return; // Don't load the preview at all if it's not needed (prevents most of the pixmaps loading late)
+                        var data = TextureData.load(previewFile(), false);
+                        data.prepare();
+                        previewQueue.add(() -> {
+                            if (preview == null) { // By the time the pixmap finished loading, we no longer needed it, so we don't create a texture.
+                                data.consumePixmap().dispose(); // Since we don't create a texture, we need to manually dispose the pixmap.
+                                return;
+                            }
+                            preview = new TextureRegion(new Texture(data));
+                        });
+                    });
+                }
             }
-            return null;
+            return preview;
+        }
+
+        public void unloadTexture(){
+            if(preview != null && preview != Core.atlas.find("nomap")) preview.texture.dispose();
+            preview = null;
         }
 
         private String index(){
@@ -251,10 +309,6 @@ public class Saves{
 
         private Fi previewFile(){
             return mapPreviewDirectory.child("save_slot_" + index() + ".png");
-        }
-
-        private Fi loadPreviewFile(){
-            return previewFile().sibling(previewFile().name() + ".spreview");
         }
 
         public boolean isHidden(){
@@ -270,7 +324,7 @@ public class Saves{
         }
 
         public String getDate(){
-            return SimpleDateFormat.getDateTimeInstance().format(new Date(meta.timestamp));
+            return dateFormat.format(new Date(meta.timestamp));
         }
 
         public Map getMap(){
@@ -332,7 +386,7 @@ public class Saves{
             try{
                 from.copyTo(file);
                 if(previewFile().exists()){
-                    requestedPreview = false;
+                    unloadTexture();
                     previewFile().delete();
                 }
             }catch(Exception e){
@@ -358,9 +412,7 @@ public class Saves{
                 current = null;
             }
 
-            if(Core.assets.isLoaded(loadPreviewFile().path())){
-                Core.assets.unload(loadPreviewFile().path());
-            }
+            unloadTexture();
         }
     }
 }
